+++ firmware/esp8266/libraries/ESPway/src/ESPwayFramework.h
@@ -14,7 +14,7 @@
 #include "ESPwayStreamOta.h"
 #include "ESPwayEnrollment.h"
 
-#define ESPWAY_FRAMEWORK_VERSION "0.4.8"
+#define ESPWAY_FRAMEWORK_VERSION "0.4.9"
 
 class ESPwayFramework {
  public: